SKMM: Add SKMM application
SKMM (Secure Key Management Module) is intended to offload public key
management tasks from host through PCIe data path or Ethernet port in
a secure way.

The host driver interfaces with OpenSSL and passes the requests to EP
using abstract request, the EP parses the abstract requests, drives
the SEC engine to do the encrypt/decrypt operations and passes the
results to host. The private key is managed on EP side and is not
visiable to host.

Here is a SKMM diagram using PCIe as data path.

           Root Complex                    End Point
	|---------------------|        |---------------------|
	|   OpenSSL           |        |  SKMM application   |
	|---------------------|        |---------------------|
	|   CryptoDev         |        |  UIO/VFIO driver    |
	|                     |  PCIe  |                     |
	|   PCIe data path    |<======>|  PCIe data path     |
	|---------------------|        |---------------------|

The SKMM application runs in user space and operates the hardware
resource (PCIe controller, SEC engine, SRAM) by UIO or VFIO. The
PCIe is used for the data path between host and EP, so for PCIe
End point demonstration, we can also use SKMM to demonstrate the
end point functionality.

Signed-off-by: Fan Zhicheng <B32736@freescale.com>
Signed-off-by: Yuanquan Chen <Yuanquan.Chen@freescale.com>
Signed-off-by: Jiucheng Xu <Jiucheng.Xu@freescale.com>
Signed-off-by: Mingkai Hu <Mingkai.hu@freescale.com>