usdpaa/dma: Update the DMA init function to a more general way
Currently the USDPAA DMA gets and initializes all DMA channels
based on the DMA DTB nodes. The function needs compatible of
the DMA controller and the property "cell-index". But in the
new DMA DTS files for some new platforms, a new compatible was
used and the "cell-index" was removed.

So it's a better way to update the DMA init function with "sysfs"
method.

Signed-off-by: Liu Gang <Gang.Liu@freescale.com>
Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>