PCIDMA: Change DMA channel initialization code
The original code initializes DMA channel in test thread. When
multiple test threads run simultaneously, there is a chance that
they initialize the same DMA channel simultaneously, this will
cause the thread to fail. The patch initialize the DMA channel at
the beginning of the process, and all test thread directly use the
initialized DMA channel.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>