fix host freeze problem due to counter mismatch
Because there is a propagation time between the device and the host some
packets may be sent by the device but not read by the host during the
service of the current interrupt. During a stream of packets this doesn't
pose a problem because the undetected packets will be read by the host
on the next interrupt. The problem arises at the end of a stream of
packets the host tries to shut down (when for example the user has
issued a CTRL+C command). The software checks if the number of received
packets is equal to the number of sent packets, and because they are not
equal gets stuck in an infinite loop.

The solution is to check, in the endpoint software, that the number of
sent packets is equal to the number of checked packets reported by the
host. If they are not equal a new interrupt is issued so that the host
may service that remaining packets.

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>