fix interrupt request format for MSI
This patch fixes a problem of incorrect interrupt request message on
Intel(R) platforms with interrupt-remapping enabled.

The error manifests as:
dmar: INTR-REMAP: Request device [[01:00.0] fault index 1e
INTR-REMAP:[fault reason 32] Detected reserved fields in the decoded interrupt-remapped request

This failure occurs because the upper 16 bits in the interrupt request
must explicitely be set to zero if interrupt-remapping is enabled.

If the upper bits are not set to zero, things work because these bits
are (or seem to be) ignored. On Intel(R) platforms with IOMMU is enabled,
this is no longer the case because Interrupt Requests are in Remappable
Format (confirmed by bit SHV of the Interrupt Address) and the upper 16
bits _must_ be set to zero or a failure mode occurs.

This patch is consistent with the PCI specification on MSI which asks
that the 16 bits Message Data register are written to the Interrupt
Address and that the upper 16 bits are set to zero.

See also:
Intel(R) Virtualization Technology for Directed I/O Architecture Specification, Rev 2.3
- 5.1.2 Interrupt Requests Format on Intel(R) 64 Platforms

PCI LOCAL BUS SPECIFICATION, Rev. 3.0
- 6.8.1.6 Message Data for MSI

Signed-off-by: Anand Singh <anand.singh@freescale.com>
Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>