Add back the implicit Makefile rules
This commit reverts a patch that was imported from usdpaa a long time ago...

The reverted patch removes builtin make variables that are essential to
proper functioning of the build process. In particular it removes the
variable "RM" which is actually used in clean target. By sheer luck, the RM
variable is not removed by this option in make < 4.0 but it is in recent
versions.
The consequence is that "clean" target fails with an obtuse message like
"make: ./include/fsl_sec/.Makefile.touch: Command not found" because
"rm" command is not substituted to the $(RM) variable which doesn't
exist.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>